Updates to new structure
