+++ package.json
@@ -55,8 +55,9 @@
     "url": "https://github.com/unmock/unmock-js/issues"
   },
   "dependencies": {
-    "unmock-core": "file:packages/unmock-core",
+    "expect": "^24.7.1",
     "unmock-cli": "file:packages/unmock-cli",
+    "unmock-core": "file:packages/unmock-core",
     "unmock-jsdom": "file:packages/unmock-jsdom",
     "unmock-node": "file:packages/unmock-node"
   },
