+++ packages/unmock-core/src/index.ts
@@ -22,6 +22,7 @@ export const util = {
   endReporter,
   hostIsWhitelisted,
 };
+export { snapshot } from "./snapshot";
 
 export const defaultOptions: IUnmockInternalOptions = {
   backend: new FailingBackend(),
